# api/openapi.yaml
openapi: 3.0.3
info:
  title: CareDesk API
  version: 1.0.0
  description: |
    REST API for the CareDesk platform.

    All resources are served under the `/api/v1` base path. Error responses follow
    RFC 9457 (Problem Details for HTTP APIs) and are returned with the
    `application/problem+json` media type.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Versioned base path
security:
  - bearerAuth: []
...
paths:
...
  /doctors:
    get:
      tags: [Doctors]
      summary: Search doctors
      operationId: listDoctors
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive search term for doctor name or specialization.
          schema:
            type: string
        - name: specialization
          in: query
          required: false
          description: Case-insensitive specialization filter.
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated doctor list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /doctors/specializations:
    get:
      tags: [Doctors]
      summary: List doctor specializations
      description: >-
        Distinct, sorted list of specializations across all enabled doctors.
        Backs the booking flow's specialization filter, which then narrows the
        doctor list via GET /doctors?specialization=. A new specialization
        appears automatically once an admin creates a doctor with it.
      operationId: listDoctorSpecializations
      responses:
        '200':
          description: Specialization names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /doctors/{doctorId}:
    get:
      tags: [Doctors]
      summary: Get doctor profile
      operationId: getDoctorById
      parameters:
        - $ref: '#/components/parameters/DoctorId'
      responses:
        '200':
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /doctors/{doctorId}/schedule:
    get:
      tags: [Doctors]
      summary: View a doctor's schedule
      operationId: getDoctorSchedule
      parameters:
        - $ref: '#/components/parameters/DoctorId'
      responses:
        '200':
          description: Doctor schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Doctors]
      summary: Add an available schedule slot
      operationId: createDoctorScheduleSlot
      parameters:
        - $ref: '#/components/parameters/DoctorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSlotCreate'
      responses:
        '201':
          description: Schedule slot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleSlot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /doctors/{doctorId}/schedule/recurring:
    post:
      tags: [Doctors]
      summary: Publish recurring availability as materialized slots
      description: >-
        Expands a weekly recurrence (weekdays, time window, slot length, date
        range) into individual schedule slots. Occurrences that overlap an
        existing slot are skipped and reported instead of failing the request.
      operationId: createDoctorRecurringSchedule
      parameters:
        - $ref: '#/components/parameters/DoctorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringScheduleCreate'
      responses:
        '201':
          description: Recurrence expanded; overlapping occurrences skipped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringScheduleResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /doctors/{doctorId}/schedule/{slotId}:
    delete:
      tags: [Doctors]
      summary: Delete an unbooked schedule slot
      operationId: deleteDoctorScheduleSlot
      parameters:
        - $ref: '#/components/parameters/DoctorId'
        - $ref: '#/components/parameters/SlotId'
      responses:
        '204':
          description: Slot deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DoctorId:
      name: doctorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SlotId:
      name: slotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Authentication is required or has failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: The caller is authenticated but not allowed to access the resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: The request conflicts with the current state of the resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: An unexpected error occurred while processing the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
    ProblemDetail:
      type: object
      description: |
        Problem Details object as defined by RFC 9457
        (https://www.rfc-editor.org/rfc/rfc9457).
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference identifying the problem type.
          default: about:blank
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence.
        errors:
          type: array
          description: Optional list of field-level validation errors.
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
          description: Zero-based page index of the returned slice.
        size:
          type: integer
          minimum: 1
          description: Number of items per page.
        totalElements:
          type: integer
          format: int64
          minimum: 0
          description: Total number of items across all pages.
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages.
    PaginatedUserProfileResponse:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
        page:
          $ref: '#/components/schemas/PageMeta'
...
    UserRole:
      type: string
      enum: [PATIENT, DOCTOR, ADMIN]
...
    UserProfile:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        clinicId:
          type: string
          format: uuid
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        specialization:
          type: string
        licenseNumber:
          type: string
        enabled:
          type: boolean
          description: Whether the account is active. Disabled accounts cannot authenticate.
        password:
          type: string
          writeOnly: true
...
    Schedule:
      type: object
      required: [doctorId, slots]
      properties:
        doctorId:
          type: string
          format: uuid
        slots:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleSlot'
    ScheduleSlot:
      type: object
      required: [id, startAt, endAt, available]
      properties:
        id:
          type: string
          format: uuid
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        available:
          type: boolean
    ScheduleSlotCreate:
      type: object
      required: [startAt, endAt]
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    Weekday:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
    RecurringScheduleCreate:
      type: object
      required: [weekdays, startTime, endTime, slotDurationMinutes, startDate, endDate, timezone]
      properties:
        weekdays:
          type: array
          items:
            $ref: '#/components/schemas/Weekday'
          minItems: 1
          uniqueItems: true
        startTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          description: Daily window start (wall clock, HH:mm).
          example: '09:00'
        endTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          description: Daily window end (wall clock, HH:mm, exclusive).
          example: '12:00'
        slotDurationMinutes:
          type: integer
          enum: [15, 30, 45, 60]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          description: Inclusive; at most 12 weeks after startDate.
        timezone:
          type: string
          description: IANA zone id used for wall-clock expansion.
          example: Europe/Berlin
    SlotInterval:
      type: object
      required: [startAt, endAt]
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    RecurringScheduleResult:
      type: object
      required: [created, skipped]
      properties:
        created:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleSlot'
        skipped:
          type: array
          description: Occurrences not created because they overlap existing slots.
          items:
            $ref: '#/components/schemas/SlotInterval'
...