# services/ai-assistant/caredesk_client/models/problem_detail.py
from __future__ import annotations

from collections.abc import Mapping
from typing import TYPE_CHECKING, Any, TypeVar

from attrs import define as _attrs_define
from attrs import field as _attrs_field

from ..types import UNSET, Unset

if TYPE_CHECKING:
    from ..models.validation_error import ValidationError


T = TypeVar("T", bound="ProblemDetail")


@_attrs_define
class ProblemDetail:
    """Problem Details object as defined by RFC 9457
    (https://www.rfc-editor.org/rfc/rfc9457).

        Attributes:
            title (str): A short, human-readable summary of the problem type.
            status (int): The HTTP status code generated by the origin server.
            type_ (str | Unset): A URI reference identifying the problem type. Default: 'about:blank'.
            detail (str | Unset): A human-readable explanation specific to this occurrence.
            instance (str | Unset): A URI reference that identifies the specific occurrence.
            errors (list[ValidationError] | Unset): Optional list of field-level validation errors.
    """

    title: str
    status: int
    type_: str | Unset = "about:blank"
    detail: str | Unset = UNSET
    instance: str | Unset = UNSET
    errors: list[ValidationError] | Unset = UNSET
    additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)

    def to_dict(self) -> dict[str, Any]:
        title = self.title

        status = self.status

        type_ = self.type_

        detail = self.detail

        instance = self.instance

        errors: list[dict[str, Any]] | Unset = UNSET
        if not isinstance(self.errors, Unset):
            errors = []
            for errors_item_data in self.errors:
                errors_item = errors_item_data.to_dict()
                errors.append(errors_item)

        field_dict: dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update(
            {
                "title": title,
                "status": status,
            }
        )
        if type_ is not UNSET:
            field_dict["type"] = type_
        if detail is not UNSET:
            field_dict["detail"] = detail
        if instance is not UNSET:
            field_dict["instance"] = instance
        if errors is not UNSET:
            field_dict["errors"] = errors

        return field_dict

    @classmethod
    def from_dict(cls: type[T], src_dict: Mapping[str, Any]) -> T:
        from ..models.validation_error import ValidationError

        d = dict(src_dict)
        title = d.pop("title")

        status = d.pop("status")

        type_ = d.pop("type", UNSET)

        detail = d.pop("detail", UNSET)

        instance = d.pop("instance", UNSET)

        _errors = d.pop("errors", UNSET)
        errors: list[ValidationError] | Unset = UNSET
        if _errors is not UNSET:
            errors = []
            for errors_item_data in _errors:
                errors_item = ValidationError.from_dict(errors_item_data)

                errors.append(errors_item)

        problem_detail = cls(
            title=title,
            status=status,
            type_=type_,
            detail=detail,
            instance=instance,
            errors=errors,
        )

        problem_detail.additional_properties = d
        return problem_detail

    @property
    def additional_keys(self) -> list[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties